GVT-2300 & GVT-2301 & GVT-2302: Add validation for splits General: Split UI components in location track splitting infobox into their own files General: Add hasErrors to infobox-field
