+++ infra/src/main/kotlin/fi/fta/geoviite/infra/tracklayout/LocationTrackDao.kt
@@ -16,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional
 import java.sql.ResultSet
 import java.sql.Timestamp
 import java.time.Instant
+import java.util.*
 
 const val LOCATIONTRACK_CACHE_SIZE = 10000L
 
@@ -300,21 +301,29 @@ class LocationTrackDao(
         publicationState: PublishType,
         includeDeleted: Boolean,
         trackNumberId: IntId<TrackLayoutTrackNumber>? = null,
-        name: AlignmentName? = null,
-    ): List<LocationTrack> = fetchVersions(publicationState, includeDeleted, trackNumberId, name).map(::fetch)
+    ): List<LocationTrack> = fetchVersions(publicationState, includeDeleted, trackNumberId).map(::fetch)
+
+    fun list(
+        publicationState: PublishType,
+        includeDeleted: Boolean,
+        trackNumberId: IntId<TrackLayoutTrackNumber>? = null,
+        names: List<AlignmentName>,
+    ): List<LocationTrack> = fetchVersions(publicationState, includeDeleted, trackNumberId, names).map(::fetch)
 
     fun fetchVersions(
         publicationState: PublishType,
         includeDeleted: Boolean,
         trackNumberId: IntId<TrackLayoutTrackNumber>? = null,
-        name: AlignmentName? = null,
+        names: List<AlignmentName> = emptyList(),
     ): List<RowVersion<LocationTrack>> {
+        val namesLowercasedOrNull = if (names.isNotEmpty()) names.map { name -> name.toString().lowercase() } else null
+
         val sql = """
             select lt.row_id, lt.row_version 
             from layout.location_track_publication_view lt
             where 
               (cast(:track_number_id as int) is null or lt.track_number_id = :track_number_id) 
-              and (cast(:name as varchar) is null or lower(lt.name) = lower(:name))
+              and (cast(:names as varchar[])) is null or lower(lt.name) in (:names)
               and :publication_state = any(lt.publication_states)
               and (:include_deleted = true or state != 'DELETED')
         """.trimIndent()
@@ -322,7 +331,7 @@ class LocationTrackDao(
             "track_number_id" to trackNumberId?.intValue,
             "publication_state" to publicationState.name,
             "include_deleted" to includeDeleted,
-            "name" to name,
+            "names" to namesLowercasedOrNull,
         )
         return jdbcTemplate.query(sql, params) { rs, _ ->
             rs.getRowVersion("row_id", "row_version")
