+++ ui/src/tool-panel/infobox/infobox-field.tsx
@@ -14,6 +14,7 @@ type InfoboxFieldProps = {
     onEdit?: () => void;
     className?: string;
     iconDisabled?: boolean;
+    hasErrors?: boolean;
 };
 
 const InfoboxField: React.FC<InfoboxFieldProps> = ({
@@ -24,6 +25,7 @@ const InfoboxField: React.FC<InfoboxFieldProps> = ({
     qaId,
     inEditMode = false,
     iconDisabled = false,
+    hasErrors = false,
     ...props
 }: InfoboxFieldProps) => {
     const classes = createClassName(styles['infobox__field'], className);
@@ -31,8 +33,20 @@ const InfoboxField: React.FC<InfoboxFieldProps> = ({
 
     return (
         <div className={classes} qa-id={qaId}>
-            <div className={styles['infobox__field-label']}>{label}</div>
-            <div className={styles['infobox__field-value']}>{children || value}</div>
+            <div
+                className={createClassName(
+                    styles['infobox__field-label'],
+                    hasErrors && styles['infobox__field-label--error'],
+                )}>
+                {label}
+            </div>
+            <div
+                className={createClassName(
+                    styles['infobox__field-value'],
+                    hasErrors && styles['infobox__field-value--error'],
+                )}>
+                {children || value}
+            </div>
             {!inEditMode && props.onEdit && !iconDisabled && (
                 <div
                     className={styles['infobox__edit-icon']}
