+++ ui/src/tool-panel/infobox/infobox.module.scss
@@ -49,6 +49,10 @@
         align-items: start;
         width: 100%;
         max-width: 120px;
+
+        &--error {
+            color: vayla-design.$color-red-dark;
+        }
     }
 
     &__field-value {
@@ -57,6 +61,14 @@
         align-items: start;
         margin-left: 24px;
         flex-grow: 1;
+
+        &--error {
+            color: vayla-design.$color-red-dark;
+
+            .infobox__text {
+                color: vayla-design.$color-red-dark;
+            }
+        }
     }
 
     &__edit-icon {
