+++ ui/src/tool-panel/location-track/location-track-infobox.scss
@@ -65,8 +65,12 @@
     min-width: 36px;
 }
 
-.location-track-infobox__split-replaces-duplicate {
+.location-track-infobox__split-remark {
     padding-top: 0px;
+}
+
+.location-track-infobox__split-error .infobox__text {
+    color: vayla-design.$color-red-dark;
 }
 
 .location-track-infobox__split-item-ball {
