+++ ui/src/track-layout/layout-location-track-api.ts
@@ -89,10 +89,10 @@ export async function getLocationTrackInfoboxExtras(
 
 export async function getLocationTracksByName(
     trackNumberId: LayoutTrackNumberId,
-    locationTrackName: string,
+    locationTrackNames: string[],
     publishType: PublishType,
 ): Promise<LayoutLocationTrack[]> {
-    const params = queryParams({ locationTrackName });
+    const params = queryParams({ locationTrackNames });
     return getNonNull<LayoutLocationTrack[]>(
         `${layoutUri('track-numbers', publishType)}/${trackNumberId}/location-tracks${params}`,
     );
